requirements.txt
# Core dependencies
numpy==2.3.4
scipy==1.16.3
matplotlib==3.10.7
seaborn==0.13.2

# Active inference
inferactively-pymdp==0.0.7.1

# Data handling
pandas==2.3.3

# Machine learning utilities
scikit-learn==1.7.2

# Progress bars
tqdm==4.67.1

arviz==0.22.0

# For behavioral data
pyarrow==22.0.0

# Testing
pytest==8.3.4